requirements.txt
streamlit
requests
PyMuPDF==1.22.5
Pillow
easyocr
torch
torchvision
numpy
